backend/requirements.txt
# Azure Weather Dashboard – Backend Dependencies
azure-functions==1.21.3
requests==2.32.3
